backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai-whisper==20230918
torch==2.1.0
torchaudio==2.1.0
ffmpeg-python==0.2.0
requests==2.31.0
deepl==1.16.1
googletrans==4.0.0rc1
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
yt-dlp==2023.11.16 